move models into common
